refactor: Add PluginActions class
