+++ src/Service/Admin/LicenseManagement/LicenseManagementManager.php
@@ -4,9 +4,9 @@ namespace WP_Statistics\Service\Admin\LicenseManagement;
 
 use Exception;
 use WP_Statistics;
+use WP_Statistics\Service\Admin\LicenseManagement\Plugin\PluginActions;
 use WP_Statistics\Service\Admin\LicenseManagement\Plugin\PluginHandler;
 use WP_Statistics\Service\Admin\LicenseManagement\Plugin\PluginUpdater;
-use WP_Statistics\Utils\Request;
 
 class LicenseManagementManager
 {
@@ -21,7 +21,7 @@ class LicenseManagementManager
 
         // Initialize the necessary components
         $this->initializeMenu();
-        $this->initializeAjaxCallbacks();
+        $this->initializeActionCallbacks();
         $this->initializePluginUpdaters();
 
         add_action('admin_init', [$this, 'checkForPluginsWithoutLicenses']);
@@ -53,139 +53,9 @@ class LicenseManagementManager
     /**
      * Initialize AJAX callbacks for various license management actions.
      */
-    private function initializeAjaxCallbacks()
+    private function initializeActionCallbacks()
     {
-        add_filter('wp_statistics_ajax_list', [$this, 'registerAjaxCallbacks']);
-    }
-
-    public function registerAjaxCallbacks($list)
-    {
-        $list[] = ['class' => $this, 'action' => 'check_license'];
-        $list[] = ['class' => $this, 'action' => 'download_plugin'];
-        $list[] = ['class' => $this, 'action' => 'check_plugin'];
-        $list[] = ['class' => $this, 'action' => 'activate_plugin'];
-
-        return $list;
-    }
-
-    public function check_license_action_callback()
-    {
-        check_ajax_referer('wp_rest', 'wps_nonce');
-
-        try {
-            $licenseKey = Request::has('license_key') ? wp_unslash(Request::get('license_key')) : false;
-
-            if (!$licenseKey) {
-                throw new Exception(__('License key is missing.', 'wp-statistics'));
-            }
-
-            // Merge the product list with the license and installation status
-            $mergedProductList = $this->apiCommunicator->mergeProductStatusWithLicense($licenseKey);
-
-            wp_send_json_success([
-                'products' => $mergedProductList,
-                'message'  => __('License is valid.', 'wp-statistics'),
-            ]);
-        } catch (Exception $e) {
-            wp_send_json_error([
-                'message' => $e->getMessage(),
-            ]);
-        }
-
-        exit;
-    }
-
-    public function download_plugin_action_callback()
-    {
-        check_ajax_referer('wp_rest', 'wps_nonce');
-
-        try {
-            $licenseKey = Request::has('license_key') ? wp_unslash(Request::get('license_key')) : false;
-            $pluginSlug = Request::has('plugin_slug') ? wp_unslash(Request::get('plugin_slug')) : false;
-
-            if (!$pluginSlug) {
-                throw new Exception(__('Plugin slug is missing.', 'wp-statistics'));
-            }
-
-            if (empty($licenseKey)) {
-                $licenseKey = $this->apiCommunicator->getValidLicenseForProduct($pluginSlug);
-            }
-
-            if (empty($licenseKey)) {
-                throw new Exception(__('License key is missing.', 'wp-statistics'));
-            }
-
-            $downloadUrl = $this->apiCommunicator->getPluginDownloadUrl($licenseKey, $pluginSlug);
-            if (!$downloadUrl) {
-                throw new Exception(__('Download URL not found!', 'wp-statistics'));
-            }
-
-            // Download and install the plugin
-            $this->pluginHandler->downloadAndInstallPlugin($downloadUrl);
-
-            wp_send_json_success([
-                'message' => __('Plugin downloaded and installed successfully.', 'wp-statistics'),
-            ]);
-        } catch (Exception $e) {
-            wp_send_json_error([
-                'message' => $e->getMessage(),
-            ]);
-        }
-
-        exit;
-    }
-
-    /**
-     * Handles `check_plugin` ajax call and returns info about a local plugin.
-     *
-     * @return void
-     */
-    public function check_plugin_action_callback()
-    {
-        check_ajax_referer('wp_rest', 'wps_nonce');
-
-        try {
-            $pluginSlug = Request::has('plugin_slug') ? wp_unslash(Request::get('plugin_slug')) : false;
-            if (!$pluginSlug) {
-                throw new Exception(__('Plugin slug is missing.', 'wp-statistics'));
-            }
-
-            wp_send_json_success([
-                'active' => $this->pluginHandler->isPluginActive($pluginSlug),
-                'data'   => $this->pluginHandler->getPluginData($pluginSlug),
-            ]);
-        } catch (Exception $e) {
-            wp_send_json_error([
-                'message' => $e->getMessage(),
-            ]);
-        }
-
-        exit;
-    }
-
-    public function activate_plugin_action_callback()
-    {
-        check_ajax_referer('wp_rest', 'wps_nonce');
-
-        try {
-            $pluginSlug = Request::has('plugin_slug') ? wp_unslash(Request::get('plugin_slug')) : false;
-
-            if (!$pluginSlug) {
-                throw new Exception(__('Plugin slug is missing.', 'wp-statistics'));
-            }
-
-            $this->pluginHandler->activatePlugin($pluginSlug);
-
-            wp_send_json_success([
-                'message' => __('Plugin activated successfully.', 'wp-statistics'),
-            ]);
-        } catch (Exception $e) {
-            wp_send_json_error([
-                'message' => $e->getMessage(),
-            ]);
-        }
-
-        exit;
+        add_filter('wp_statistics_ajax_list', [new PluginActions(), 'registerAjaxCallbacks']);
     }
 
     /**
